Fix tick variable types (core branch).

+++ linux/hashdialog.hh
@@ -38,17 +38,17 @@ class Hash:
 
 	private:
 		// GUI functions
-		void updateStats_gui(std::string file, int64_t bytes, size_t files, uint32_t tick);
+		void updateStats_gui(std::string file, int64_t bytes, size_t files, uint64_t tick);
 
 		// Client callbacks
-		virtual void on(dcpp::TimerManagerListener::Second, uint32_t tics) throw();
+		virtual void on(dcpp::TimerManagerListener::Second, uint64_t tics) throw();
 
 		// GUI callback
 		static void onPauseHashing_gui(GtkWidget *widget, gpointer data);//NOTE: core 0.762
 
 		int64_t startBytes;
 		size_t startFiles;
-		uint32_t startTime;
+		uint64_t startTime;
 };
 
 #else
